files

+++ controllers/admin-mg.js
@@ -1,3 +1,4 @@
+const fileHelper = require('../utils/file')
 const Product = require('../models/product-mg')
 
 exports.getAddProduct = (req, res, next) => {
@@ -9,7 +10,27 @@ exports.getAddProduct = (req, res, next) => {
 }
 
 exports.postAddProduct = (req, res, next) => {
-  const {title, imageUrl, price, description} = req.body
+  const {title, price, description} = req.body
+  const {file: image} = req
+  
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+    })
+  }
+
+  const imageUrl = image.path
+
   const product = new Product({
     title,
     price,
@@ -54,7 +75,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTitle = req.body.title
   const updatedPrice = req.body.price
-  const updatedImageUrl = req.body.imageUrl
+  const image = req.file
   const updatedDescription = req.body.description
 
   Product
@@ -66,7 +87,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle
       product.price = updatedPrice
       product.description = updatedDescription
-      product.imageUrl = updatedImageUrl
+      
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
 
       return product
         .save()
@@ -93,12 +118,23 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId
-  Product
-    // .findByIdAndRemove(prodId)
-    .deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then(prod => {
+      if (!prod) {
+        return next(new Error('Product not found'))
+      }
+      fileHelper.deleteFile(prod.imageUrl)
+      return Product
+        // .findByIdAndRemove(prodId)
+        .deleteOne({_id: prodId, userId: req.user._id})
+    })
     .then(() => {
       console.log('Destroyed Product')
       res.redirect('/admin/products')
     })
-    .catch(e => console.log({e}))
+    .catch(e => {
+      const error = new Error(e)
+      error.httpStatusCode = 500
+      return next(error)
+    })
 }
